:recycle: Create `cloud` module
