+++ cloud/bots.go
@@ -17,11 +17,11 @@
 // note: this file is generated by ../scripts/generate_bots.go and uses data from:
 // https://github.com/JetBrains/qodana-cli/blob/main/bots.json
 
-package core
+package cloud
 
 var (
-	gitHubBotSuffix = "[bot]@users.noreply.github.com"
-	commonGitBots   = []string{
+	GitHubBotSuffix = "[bot]@users.noreply.github.com"
+	CommonGitBots   = []string{
 		"cla-bot@users.noreply.github.com",
 		"codecov-io@users.noreply.github.com",
 		"snyk-bot@snyk.io",
