+++ cloud/license_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package core
+package cloud
 
 import (
 	"fmt"
@@ -29,8 +29,8 @@ import (
 func TestRequestLicenseData(t *testing.T) {
 	expectedLicense := "license data"
 	rightToken := "token data"
-	os.Setenv(qodanaLicenseRequestCooldownEnv, "2")
-	os.Setenv(qodanaLicenseRequestTimeoutEnv, "6")
+	os.Setenv(QodanaLicenseRequestCooldownEnv, "2")
+	os.Setenv(QodanaLicenseRequestTimeoutEnv, "6")
 	for _, testData := range []struct {
 		name           string
 		delay          int
@@ -118,7 +118,7 @@ func TestRequestLicenseData(t *testing.T) {
 			}))
 			defer svr.Close()
 
-			res, err := requestLicenseData(svr.URL, testData.token)
+			res, err := RequestLicenseData(svr.URL, testData.token)
 			if err != nil {
 				if testData.success {
 					t.Errorf("requestLicenseData should failed")
@@ -172,122 +172,10 @@ func TestExtractLicenseKey(t *testing.T) {
 		},
 	} {
 		t.Run(testData.name, func(t *testing.T) {
-			key := extractLicenseKey([]byte(testData.data))
+			key := ExtractLicenseKey([]byte(testData.data))
 			if key != testData.expectedKey {
 				t.Errorf("expected key to be '%s' got '%s'", key, testData.expectedKey)
 			}
 		})
-	}
-}
-
-func TestSetupLicenseToken(t *testing.T) {
-	for _, testData := range []struct {
-		name       string
-		token      string
-		loToken    string
-		resToken   string
-		sendFus    bool
-		sendReport bool
-	}{
-		{
-			name:       "no key",
-			token:      "",
-			loToken:    "",
-			resToken:   "",
-			sendFus:    true,
-			sendReport: false,
-		},
-		{
-			name:       "with token",
-			token:      "a",
-			loToken:    "",
-			resToken:   "a",
-			sendFus:    true,
-			sendReport: true,
-		},
-		{
-			name:       "with license only token",
-			token:      "",
-			loToken:    "b",
-			resToken:   "b",
-			sendFus:    false,
-			sendReport: false,
-		},
-		{
-			name:       "both tokens",
-			token:      "a",
-			loToken:    "b",
-			resToken:   "a",
-			sendFus:    true,
-			sendReport: true,
-		},
-	} {
-		t.Run(testData.name, func(t *testing.T) {
-			err := os.Setenv(qodanaLicenseOnlyToken, testData.loToken)
-			if err != nil {
-				t.Fatal(err)
-			}
-			err = os.Setenv(qodanaToken, testData.token)
-			if err != nil {
-				t.Fatal(err)
-			}
-			setupLicenseToken(&QodanaOptions{})
-
-			if licenseToken.Token != testData.resToken {
-				t.Errorf("expected token to be '%s' got '%s'", testData.resToken, licenseToken.Token)
-			}
-
-			sendFUS := licenseToken.isAllowedToSendFUS()
-			if sendFUS != testData.sendFus {
-				t.Errorf("expected allow FUS to be '%t' got '%t'", testData.sendFus, sendFUS)
-			}
-
-			toSendReports := licenseToken.isAllowedToSendReports()
-			if toSendReports != testData.sendReport {
-				t.Errorf("expected allow send report to be '%t' got '%t'", testData.sendReport, toSendReports)
-			}
-
-			err = os.Unsetenv(qodanaLicenseOnlyToken)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			err = os.Unsetenv(qodanaToken)
-			if err != nil {
-				t.Fatal(err)
-			}
-		})
-	}
-}
-
-func TestSetupLicense(t *testing.T) {
-	Prod.Code = "QDJVM"
-	Prod.EAP = false
-	license := `{"licenseId":"VA5HGQWQH6","licenseKey":"VA5HGQWQH6","expirationDate":"2023-07-31","licensePlan":"EAP_ULTIMATE_PLUS"}`
-	expectedKey := "VA5HGQWQH6"
-
-	svr := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		_, _ = fmt.Fprint(w, license)
-	}))
-	defer svr.Close()
-	err := os.Setenv(qodanaLicenseEndpoint, svr.URL)
-	if err != nil {
-		t.Fatal(err)
-	}
-	setupLicense("token")
-
-	licenseKey := os.Getenv(qodanaLicense)
-	if licenseKey != expectedKey {
-		t.Errorf("expected key to be '%s' got '%s'", expectedKey, licenseKey)
-	}
-
-	err = os.Unsetenv(qodanaLicenseEndpoint)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	err = os.Unsetenv(qodanaLicense)
-	if err != nil {
-		t.Fatal(err)
 	}
 }
